backend/pyproject.toml
[project]
name = "knowledge-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi>=0.115.8",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
